backend/scripts/populate_cuisines/requirements.txt
psycopg2
python_dotenv
